refactor(skills): declare SkillLoadError once in shared-types
Addresses review item 2 on PR #480. The type was defined three times with
two different shapes: exported from `skills/registry.ts` (a logic file,
without `agentId`), duplicated in `apps/web/src/lib/types.ts` (with
`agentId`), and written inline twice more in `routes/skills.ts`. AGENTS.md
requires shared types to live in `packages/shared-types` and never be
exported from logic files.

Declares it once in `@openaidy/shared-types` (new `skills.ts`) with
`agentId` optional — the shape the route already produced and the web
already consumed — and has the registry, the route, and the web client
import it from there. `skills/index.ts` no longer re-exports it; nothing
imported it from that barrel. Web keeps a `./types` re-export so existing
import sites are untouched.
