+++ apps/server/src/skills/index.ts
@@ -3,11 +3,9 @@
  */
 
 export { SkillRegistry, createSkillRegistry } from './registry.js';
-export type {
-  SkillRegistryOptions,
-  SkillSummary,
-  SkillLoadError,
-} from './registry.js';
+// `SkillLoadError` is deliberately not re-exported here — it is part of the
+// shared contract and must be imported from `@openaidy/shared-types`.
+export type { SkillRegistryOptions, SkillSummary } from './registry.js';
 export { parseSkillMd } from './parser.js';
 export type { SkillDefinition, SkillParseError } from './parser.js';
 export {
